+++ XmlPresetBackups/ReachUIData/ReachDesktop.xml
@@ -286,25 +286,25 @@
                        allowDragging="1" allowCallbacks="Clicks, Hover &amp; Dragging"
                        itemColour="4278239743" text="Reverb"/>
             <Component type="ScriptPanel" id="panel_effectTab2" x="96.90000152587891"
-                       y="0" width="99.0" height="40.0" parentComponent="panel_customizerHeader"
+                       y="0.0" width="99.0" height="40.0" parentComponent="panel_customizerHeader"
                        allowDragging="1" allowCallbacks="Clicks, Hover &amp; Dragging"
                        itemColour="4278239743" text="Degrade"/>
             <Component type="ScriptPanel" id="panel_effectTab3" x="194.1000061035156"
-                       y="0" width="99.0" height="40.0" parentComponent="panel_customizerHeader"
+                       y="0.0" width="99.0" height="40.0" parentComponent="panel_customizerHeader"
                        allowDragging="1" allowCallbacks="Clicks, Hover &amp; Dragging"
                        itemColour="4278239743" text="Flanger"/>
             <Component type="ScriptPanel" id="panel_effectTab4" x="291.2999877929688"
-                       y="0" parentComponent="panel_customizerHeader" width="99.0" height="40.0"
-                       allowDragging="1" allowCallbacks="Clicks, Hover &amp; Dragging"
+                       y="0.0" parentComponent="panel_customizerHeader" width="99.0"
+                       height="40.0" allowDragging="1" allowCallbacks="Clicks, Hover &amp; Dragging"
                        itemColour="4278239743" text="Chorus"/>
             <Component type="ScriptPanel" id="panel_effectTab5" x="387.2999877929688"
-                       y="0" width="99.0" height="40.0" parentComponent="panel_customizerHeader"
+                       y="0.0" width="99.0" height="40.0" parentComponent="panel_customizerHeader"
                        allowDragging="1" allowCallbacks="Clicks, Hover &amp; Dragging"
                        itemColour="4278239743" text="Distort"/>
           </Component>
           <Component type="ScriptPanel" id="effectDisplay-Reverb" x="0.0" y="40.0"
                      width="490.0" height="190.0" parentComponent="panel_effectCustomizer"
-                     textColour="16777215" visible="0">
+                     textColour="16777215">
             <Component type="ScriptSlider" id="displayKnob_reverb_feedback" x="60.0"
                        y="60.0" width="70.0" height="110.0" parentComponent="effectDisplay-Reverb"
                        text="FEEDBACK" mode="NormalizedPercentage" suffix="%" middlePosition="0.5"
@@ -356,7 +356,7 @@
           <Component type="ScriptButton" id="displayButton_reverb_bypass" x="10.0"
                      y="60.0" width="12.0" height="12.0" text="displayIcon-onOff"
                      parentComponent="panel_effectCustomizer" isPluginParameter="1"
-                     pluginParameterName="Reverb Bypass" visible="0"/>
+                     pluginParameterName="Reverb Bypass"/>
           <Component type="ScriptPanel" id="effectDisplay-Degrade" x="0.0" y="40.0"
                      width="490.0" height="200.0" parentComponent="panel_effectCustomizer"
                      textColour="16777215" visible="0">
@@ -490,7 +490,7 @@
                      pluginParameterName="Distortion Bypass" visible="0"/>
           <Component type="ScriptPanel" id="effectDisplay-Chorus" x="0.0" y="40.0"
                      width="490.0" height="160.0" parentComponent="panel_effectCustomizer"
-                     textColour="16777215">
+                     textColour="16777215" visible="0">
             <Component type="ScriptPanel" id="displayPanel_chorusIndicator" x="3.0"
                        y="0.0" parentComponent="effectDisplay-Chorus" width="480.0"
                        height="4.0" itemColour="4286244807" itemColour2="4286244807"
@@ -516,7 +516,7 @@
           <Component type="ScriptButton" id="displayButton_chorus_bypass" x="10.0"
                      y="60.0" width="12.0" height="12.0" text="displayIcon-onOff"
                      parentComponent="panel_effectCustomizer" isPluginParameter="1"
-                     pluginParameterName="Chorus Bypass"/>
+                     pluginParameterName="Chorus Bypass" visible="0"/>
         </Component>
       </Component>
       <Component type="ScriptPanel" id="display_filter" x="-2.0" y="8.0" width="500.0"
@@ -610,22 +610,26 @@
             <Component type="ScriptPanel" id="displayPanel_serial" x="0.0" y="-5.0"
                        width="500.0" height="250.0" parentComponent="dis-settings_activate"
                        itemColour="16711680" itemColour2="0" bgColour="16777215" textColour="0">
-              <Component type="ScriptLabel" id="serial_error_label" x="55.0" y="150.0"
+              <Component type="ScriptLabel" id="serial_error_label" x="-5.0" y="210.0"
                          width="385.0" height="35.0" textColour="4293606219" parentComponent="displayPanel_serial"
-                         alignment="left" fontName="jetbrains-mono" text="" fontSize="18.0"/>
-              <Component type="ScriptButton" id="button_buy_reach" x="325.0" y="200.0"
-                         width="175.0" height="45.0" text="displayButton-Buy REACH" parentComponent="displayPanel_serial"
+                         alignment="left" fontName="inter-reg" text="" fontSize="18.0"/>
+              <Component type="ScriptButton" id="button_buy_reach" x="385.0" y="200.0"
+                         width="115.0" height="45.0" text="displayButton-Buy REACH" parentComponent="displayPanel_serial"
                          bgColour="4294967295" textColour="4278190080" itemColour="4292532954"
                          isMomentary="1" saveInPreset="0" defaultValue="1.0" enableMidiLearn="0"/>
-              <Component type="ScriptPanel" id="Panel2" x="55.0" y="45.0" width="385.0"
+              <Component type="ScriptButton" id="displayButton-paste" x="180.0" y="25.0"
+                         parentComponent="displayPanel_serial" width="130.0" height="35.0"
+                         text="displayButton-PASTE" saveInPreset="0" isVertical="0" enableMidiLearn="0"
+                         isMomentary="1"/>
+              <Component type="ScriptPanel" id="Panel2" x="55.0" y="75.0" width="385.0"
                          height="45.0" textColour="4286940549" borderRadius="2.0" itemColour2="0"
                          itemColour="0" parentComponent="displayPanel_serial">
                 <Component type="ScriptLabel" id="displayLabel_serialKey" x="5.0" y="10.0"
-                           width="375.0" height="25.0" fontSize="32.0" fontName="jetbrains-mono"
+                           width="375.0" height="25.0" fontSize="18.0" fontName="jetbrains-mono"
                            alignment="left" parentComponent="Panel2" text="" locked="1"/>
               </Component>
               <Component type="ScriptButton" id="displayButton_activateSerial" x="55.0"
-                         y="100.0" width="385.0" height="45.0" text="ACTIVATE" parentComponent="displayPanel_serial"
+                         y="130.0" width="385.0" height="45.0" text="ACTIVATE" parentComponent="displayPanel_serial"
                          saveInPreset="0" enableMidiLearn="0" isMomentary="1"/>
             </Component>
           </Component>
