+++ project_info.xml
@@ -2,7 +2,7 @@
 
 <ProjectSettings>
   <Name value="Reach"/>
-  <Version value="1.6.0"/>
+  <Version value="1.6.1"/>
   <Description value="Spaced Out Reverb"/>
   <BundleIdentifier value="com.sinuslabs.reach"/>
   <PluginCode value="Reac"/>
